+++ src/audio.h
@@ -179,7 +179,7 @@ extern std::int16_t LASTMUSIC;
 enum class AudioDriverType
 {
 	auto_detect,
-	r2_sdl,
+	r2_system,
 	r3_openal,
 }; // AudioDriverType