+++ src/bstone_globals.cpp
@@ -15,7 +15,6 @@ CCmdMgrUPtr ccmd_mgr{};
 PageMgrUPtr page_mgr{};
 
 sys::SystemMgrUPtr sys_system_mgr{};
-sys::AudioMgrUPtr sys_audio_mgr{};
 sys::EventMgrUPtr sys_event_mgr{};
 sys::VideoMgrUPtr sys_video_mgr{};
 sys::MouseMgrUPtr sys_mouse_mgr{};