+++ src/bstone_globals.h
@@ -8,6 +8,7 @@ SPDX-License-Identifier: GPL-2.0-or-later
 #include "bstone_ccmd_mgr.h"
 #include "bstone_cvar_mgr.h"
 #include "bstone_page_mgr.h"
+#include "bstone_sys_push_audio_device.h"
 #include "bstone_sys_system_mgr.h"
 
 namespace bstone {
@@ -22,7 +23,6 @@ extern CCmdMgrUPtr ccmd_mgr;
 extern PageMgrUPtr page_mgr;
 
 extern sys::SystemMgrUPtr sys_system_mgr;
-extern sys::AudioMgrUPtr sys_audio_mgr;
 extern sys::EventMgrUPtr sys_event_mgr;
 extern sys::VideoMgrUPtr sys_video_mgr;
 extern sys::MouseMgrUPtr sys_mouse_mgr;